changed folder names (directory?) from c1 to chapter1 etc.
need to better understand and use git to figure out how to do this
without creating n* new files and deleting n* old files. this seems to
be a bit off, and well i'm hoping that i can look back at old commits
w/ code that i commited that are in new locations?